# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Restaurant Reservation API",
        "description": "API para gestionar reservas de restaurante",
        "contact": {
            "name": "API Support",
            "email": "admin@example.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "API Server"
        }
    ],
    "paths": {
        "/api/customers": {
            "get": {
                "tags": [
                    "Customers"
                ],
                "summary": "Get all customers",
                "operationId": "4a3eaa79eb8ed63a7084b6e2c69c4b1e",
                "responses": {
                    "200": {
                        "description": "List of all customers",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Customer"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Customers"
                ],
                "summary": "Create a new customer",
                "operationId": "19f0f44c8d376351f98af50c827f9168",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nombre": {
                                        "type": "string",
                                        "example": "Juan Pérez"
                                    },
                                    "correo": {
                                        "type": "string",
                                        "example": "juan.perez@example.com"
                                    },
                                    "telefono": {
                                        "type": "string",
                                        "example": "555-123-4567"
                                    },
                                    "direccion": {
                                        "type": "string",
                                        "example": "Av. Insurgentes Sur 123, CDMX"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Customer created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Customer"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/api/customers/{id}": {
            "get": {
                "tags": [
                    "Customers"
                ],
                "summary": "Get a customer by ID",
                "operationId": "eb4260716829341d88bffd5d77c52c73",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the customer",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Customer details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Customer"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Customer not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Customers"
                ],
                "summary": "Update a customer",
                "operationId": "d8f739047e264cb6a201ac6d6c7cb36a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the customer to update",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nombre": {
                                        "type": "string",
                                        "example": "Juan Pérez Actualizado"
                                    },
                                    "correo": {
                                        "type": "string",
                                        "example": "juan.perez.update@example.com"
                                    },
                                    "telefono": {
                                        "type": "string",
                                        "example": "555-987-6543"
                                    },
                                    "direccion": {
                                        "type": "string",
                                        "example": "Nueva Dirección 456, CDMX"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Customer updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Customer"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Customer not found"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Customers"
                ],
                "summary": "Delete a customer",
                "operationId": "18ae4527e24f8dcdf3ba06d604560734",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the customer to delete",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Customer deleted successfully"
                    },
                    "404": {
                        "description": "Customer not found"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Customer": {
                "title": "Customer Model",
                "description": "Customer model representing a restaurant customer",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "nombre": {
                        "type": "string",
                        "example": "Juan Pérez"
                    },
                    "correo": {
                        "type": "string",
                        "format": "email",
                        "example": "juan.perez@example.com"
                    },
                    "telefono": {
                        "type": "string",
                        "example": "555-123-4567"
                    },
                    "direccion": {
                        "type": "string",
                        "example": "Av. Insurgentes Sur 123, CDMX"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2023-06-15T19:30:00.000000Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2023-06-15T19:30:00.000000Z"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}